Move AppDbContext references to Infrastructure, and remove Infrastructure dependency from Web.
