+++ src/CleanArchitecture.Infrastructure/CleanArchitecture.Infrastructure.csproj
@@ -17,4 +17,8 @@
     </PackageReference>
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.1.4" />
   </ItemGroup>
+
+  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
+    <Exec Command="copy /Y &quot;$(TargetDir)$(TargetName).dll&quot; &quot;$(SolutionDir)src\CleanArchitecture.Web\$(OutDir)\$(TargetName).dll&quot;" />
+  </Target>
 </Project>
