+++ src/CleanArchitecture.Web/IDatabaseRegistrar.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace CleanArchitecture.Web
+{
+    public interface IDatabaseRegistrar
+    {
+        void RegisterDb(IServiceCollection services, string connectionString);
+    }
+}
